Update iOS project to reflect new library structure
