+++ complex objects/ios/ComplexRust/ComplexRust-Bridging-Header.h
@@ -5,6 +5,6 @@
 #ifndef ComplexRust_Bridging_Header_h
 #define ComplexRust_Bridging_Header_h
 
-#import "db.h"
+#import "toodle.h"
 
 #endif /* ComplexRust_Bridging_Header_h */
