+++ complex objects/ios/ComplexRust/Rust/List/Categories/Category.swift
