+++ complex objects/ios/ComplexRust/Rust/Logins/LoginManager.swift
